Refactor package name - String and StringBuilder.
